--- system/tests/__init__.py
@@ -1,3 +1,0 @@
-from tools.lib.tests import temporary_mock_dir
-
-temporary_swaglog_dir = temporary_mock_dir("openpilot.system.swaglog.SWAGLOG_DIR")
