--- tools/lib/tests/__init__.py
@@ -1,17 +1,0 @@
-import tempfile
-
-from unittest import mock
-
-
-def temporary_mock_dir(mock_path):
-  def decorator(func):
-    def wrapper(*args, **kwargs):
-      with tempfile.TemporaryDirectory() as temp_dir:
-        cache_dir_patch = mock.patch(mock_path, temp_dir)
-        cache_dir_patch.start()
-        func(*args, **kwargs, temp_dir=temp_dir)
-        cache_dir_patch.stop()
-      return wrapper
-  return decorator
-
-temporary_cache_dir = temporary_mock_dir("openpilot.tools.lib.url_file.CACHE_DIR")
