+++ tools/lib/tests/test_caching.py
@@ -2,7 +2,7 @@
 import os
 import unittest
 from openpilot.tools.lib.url_file import URLFile
-from openpilot.tools.lib.tests import temporary_cache_dir
+from openpilot.common.mock import temporary_cache_dir
 
 
 class TestFileDownload(unittest.TestCase):
